Data.Queue.Binomial

Description

Simple binomial heaps, with a statically-enforced shape.

Documentation

data Binomial rk a

A binomial heap, where the sizes of the nodes are enforced in the types.

The implementation is based on:

Hinze, Ralf. “Functional Pearls: Explaining Binomial Heaps.” Journal of Functional Programming 9, no. 1 (January 1999): 93–104. doi:10.1017/S0956796899003317.

Wasserman, Louis. “Playing with Priority Queues.” The Monad.Reader, May 12, 2010.

It is a list of binomial trees, equivalent to a binary number (stored least-significant-bit first).

Constructors

Nil
The empty heap

Skip (Binomial (S rk) a)
Skip a child tree (equivalent to a zero in the binary representation of the data structure).

(:-) !(Tree rk a) (Binomial (S rk) a) infixr 5
A child tree. Equivalent to a one in the binary representation.

data Node n a where

A list of binomial trees, indexed by their sizes in ascending order.

Constructors

NilN :: Node Z a

(:<) :: !(Tree n a) -> Node n a -> Node (S n) a

data Tree rk a

A rose tree, where the children are indexed.

Constructors

Root a (Node rk a)