Data.Queue.Braun

Description

Simple, unchecked braun heaps.

Documentation

newtype Braun a

A Braun heap. Based on this implementation.

A braun tree is a nearly balanced binary tree: the left branch can be either exactly the same size as the right, or one element larger.

This version is unchecked (very unchecked), and is provided mainly for comparison to the checked version.

Constructors

Braun

Fields

runBraun :: Tree a