Data.Queue.Class

Description

Classes for the various heaps, mainly to avoid name clashing.

Documentation

class Queue h a where

A class for queues. Conforming members can have their own definition of order on their contents. (i.e., Ord is not required)

Minimal complete definition

minView, insert, empty

Methods

minView :: h a -> Maybe (a, h a)

Return the first element, and the remaining elements, or Nothing if the queue is empty. For most queues, this will be the minimal element

insert :: a -> h a -> h a

Insert an element into the queue.

empty :: h a

The empty queue.

singleton :: a -> h a

A queue with one element.

toList :: h a -> [a]

Return a list of the contents of the queue, in order, from smallest to largest.

fromList :: [a] -> h a

Create a heap from a list.

heapSort :: p h -> [a] -> [a]

Perform heap sort on a list of items.

class Queue h a => MeldableQueue h a where

A class for meldable queues. Conforming members should form a monoid under merge and empty.

Minimal complete definition

merge

Methods

merge :: h a -> h a -> h a

Merge two heaps. This operation is associative, and has the identity of empty.

merge x (merge y z) = merge (merge x y) z

merge x empty = merge empty x = x

fromFoldable :: Foldable f => f a -> h a

Create a heap from a Foldable container. This operation is provided to allow the use of foldMap, which may be asymptotically more efficient. The default definition uses foldMap.

showsPrecQueue :: (Queue h a, Show a) => Int -> h a -> ShowS

A default definition for showsPrec.

readPrecQueue :: (Read a, Queue h a) => Int -> ReadS (h a)

A default definition for readsPrec.

eqQueue :: (Eq a, Queue h a) => h a -> h a -> Bool

A default definition of ==.

cmpQueue :: (Ord a, Queue h a) => h a -> h a -> Ordering

A default definition of compare.