Data.Queue.Indexed.Erased

Description

Erase the size parameter on a size-indexed heap, using existentials.

Documentation

data ErasedSize f a

This type contains a size-indexed heap, however the size index is hidden. This allows it to act like a standard heap, while maintaining the proven invariants of the size-indexed version.

Constructors

ErasedSize

Fields

runErasedSize :: f n a