Data.Queue.WithDict

Description

Provides a wrapper for queues, allowing them to conform to Foldable.

Documentation

data WithDict f a where

This stores the dictionary of methods for the priority queue of f, allowing the entire type to conform to Foldable.

Constructors

WithDict :: Queue f a => f a -> WithDict f a